+++ apps/provisioning/pkg/repository/repository.go
@@ -189,6 +189,22 @@ type WebhookSecretRotator interface {
 	RotateWebhookSecret(ctx context.Context) ([]map[string]any, error)
 }
 
+type WebhookClient interface {
+	CreateWebhook(ctx context.Context, hook Webhook) (Webhook, error)
+	GetWebhook(ctx context.Context, webhookID int64) (Webhook, error)
+	EditWebhook(ctx context.Context, hook Webhook) error
+	DeleteWebhook(ctx context.Context, webhookID int64) error
+}
+
+type PullRequestClient interface {
+	CreatePullRequestComment(ctx context.Context, prNumber int, body string) error
+}
+
+type jointClient interface {
+	WebhookClient
+	PullRequestClient
+}
+
 type FileAction string
 
 const (
